requirements.txt
apscheduler
amqp==2.5.2
aniso8601==8.0.0
billiard==3.6.3.0
blinker==1.4
bson==0.5.9
celery==4.4.1
click==7.1.1
Flask==1.1.1
Flask-Cors==3.0.8
Flask-Login==0.5.0
Flask-Principal==0.4.0
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
gevent==1.4.0
greenlet==0.4.15
gunicorn==20.0.4
importlib-metadata==1.5.0
itsdangerous==1.1.0
Jinja2==2.11.1
kombu==4.6.8
MarkupSafe==1.1.1
passlib==1.7.2
pymongo==3.10.1
python-dateutil==2.8.1
python-dotenv==0.12.0
pytz==2019.3
six==1.14.0
SQLAlchemy==1.3.13
vine==1.3.0
Werkzeug==1.0.0
zipp==3.1.0
